requirements.txt
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
comm==0.2.1
confection==0.1.4
contourpy==1.2.0cf
cycler==0.12.1
cymem==2.0.8
DataProperty==1.0.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
emoji==2.10.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl#sha256=61e6e5530941f5880166855f09f60d7e6ba79ec1e8e45f96244bdb1eb169eb1d
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.49.0
fqdn==1.5.1
fsspec==2024.2.0
gensim==4.3.2
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==3.10.1
ipykernel==6.29.2
ipython==8.21.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
langcodes==3.3.0
Levenshtein==0.25.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mbstrdecoder==1.1.3
mistune==3.0.2
mpmath==1.3.0
murmurhash==1.0.10
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest_asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.1.0
notebook_shim==0.2.4
numpy==1.26.4
overrides==7.7.0
packaging==23.2
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.3
pathlib_abc==0.1.1
pathvalidate==3.2.0
pathy==0.11.0
pcfg==0.1.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==4.2.0
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.42
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.8.2
pydantic_core==2.16.2
PYEVALB==0.1.3
Pygments==2.17.2
pyparsing==3.1.1
pytablewriter==1.2.0
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-json-logger==2.0.7
python-Levenshtein==0.25.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.6.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
Send2Trash==1.8.2
setuptools==68.2.2
six==1.16.0
sklearn-crfsuite==0.3.6
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.4
spacy-conll==3.4.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-stanza==1.0.4
srsly==2.4.8
stack-data==0.6.2
stanza==1.6.1
sympy==1.12
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.4
terminado==0.18.0
thinc==8.2.3
threadpoolctl==3.3.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
torch==2.2.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.0
typepy==1.3.2
typer==0.4.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
wasabi==0.10.1
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
wheel==0.41.2
widgetsnbextension==4.0.10
zipp==3.17.0
